Lezione_10/portfolio-main/requirements.txt
flask
Flask-SQLAlchemy
pymysql